Start building a better deployment system

+++ go/build.sh
@@ -6,15 +6,16 @@ export GOPATH=`pwd`
 cd src
 for f in */ ; do
    cd $f
-   echo "Checking formatting in ${f}"
-   count=`git ls-files | grep '.go$' | xargs gofmt -l | wc -l`
-   if [ $count -gt 0 ]; then
-       echo "Some files are not formatted. run prettify.sh\n"
-       exit 1
+   if `ls -l *.go | wc -l`>0; then
+       echo "Checking formatting in ${f}"
+       count=`git ls-files | grep '.go$' | xargs gofmt -l | wc -l`
+       if [ $count -gt 0 ]; then
+           echo "Some files are not formatted. run prettify.sh\n"
+           exit 1
+       fi
+       echo "Building ${f}"
+       go test -v
+       go install -v
    fi
-   echo "Building ${f}"
-   go test -v
-   go install -v
-   echo "Done"
    cd ..
 done
